+++ packages/Accordion/src/Accordion.tsx
@@ -14,27 +14,33 @@ interface IAccordionProps {
 
 const Accordion = ({ children, isOpen, onOpen, onClose }: IAccordionProps): JSX.Element => {
   const [accordionStatus, setAccordionStatus] = useState<string>(isOpen ? FLEX : NONE);
-
   const [header, content]: [JSX.Element, JSX.Element] = children;
+  const isActive = accordionStatus === FLEX;
 
   useEffect(() => {
     setAccordionStatus(isOpen ? FLEX : NONE);
   }, [isOpen]);
 
-  const isActive: boolean = accordionStatus === FLEX;
+  useEffect(() => {
+    if (accordionStatus === FLEX) {
+      onOpen && onOpen();
+    } else {
+      onClose && onClose();
+    }
+  }, [accordionStatus, onClose, onOpen]);
+
   const toggleAccordion = (): void => {
-    setAccordionStatus(isActive ? NONE : FLEX);
-    isOpen && onOpen();
-    !isOpen && onClose();
+    const status = isActive ? NONE : FLEX;
+    setAccordionStatus(status);
   };
 
   const arrowType = isActive ? UP : DOWN;
 
   return (
-    <div className="accordion">
+    <div className="accordion" onClick={toggleAccordion} data-testid="toggle-accordion">
       <div className="accordion__header">
         {header}
-        <div className="icon-button icon-button--accordion" data-testid="toggle-accordion" onClick={toggleAccordion}>
+        <div className="icon-button icon-button--accordion">
           <i className="material-icons">{`keyboard_arrow_${arrowType}`}</i>
         </div>
       </div>