+++ packages/FileUpload/__tests__/FileInputs/FileInputs.test.tsx
@@ -5,6 +5,7 @@ import React from 'react';
 
 const handleSelectChange = jest.fn();
 const handleInputChange = jest.fn();
+const copyToOthers = jest.fn();
 
 const mockedProps = {
   data: { altText: '', caption: '' },
@@ -14,6 +15,7 @@ const mockedProps = {
     name: 'test-image-name',
     type: 'image',
   },
+  copyToOthers: copyToOthers,
   handleInputChange: handleInputChange,
   handleSelectChange: handleSelectChange,
   index: 0,
@@ -33,14 +35,29 @@ describe('FileInputs component', () => {
   it('should call mocked function handleInputChange & change the input value', () => {
     render(<FileInputs {...mockedProps} />);
     const mockedValue = 'typing first input';
-    fireEvent.change(screen.getByLabelText('Url'), { target: { value: mockedValue } });
+    fireEvent.change(screen.getByLabelText('image title'), { target: { value: mockedValue } });
     expect(handleInputChange).toHaveBeenCalled();
+  });
+
+  it('should call copyToOthers', () => {
+    const { getByTestId } = render(<FileInputs {...mockedProps} />);
+    const mockedValue = 'typing first input';
+    fireEvent.click(getByTestId('copy-to-storeKey'));
+    fireEvent.click(getByTestId('copy-to-title'));
+    expect(copyToOthers).toHaveBeenCalledWith('storeKey', mockedProps.index);
+    expect(copyToOthers).toHaveBeenCalledWith('title', mockedProps.index);
   });
 
   it('should call mocked function handleSelectChange & change the select value', () => {
     const mockedValue = 'Test';
     render(<FileInputs {...mockedProps} />);
     fireEvent.change(screen.getByLabelText('Alternative text'), { target: { value: mockedValue } });
+  });
+
+  it('should render childs based on type (2 inputs if doc)', () => {
+    const updatedProps = { ...mockedProps };
+    updatedProps.file.type = 'doc';
+    render(<FileInputs {...updatedProps} />);
   });
 
   it('should render childs based on type (2 inputs if doc)', () => {