+++ packages/FileUpload/src/FileList/index.tsx
@@ -2,18 +2,19 @@ import React, { Fragment } from 'react';
 import { DOC, IMAGE } from '../constants';
 import DocumentIcon from '../DocumentIcon';
 import FileInputs from '../FileInputs';
-
 const FileList = ({
   previewImages,
   handleCancel,
   handleInputChange,
   handleSelectChange,
   selectOptions,
+  copyToOthers
 }: {
   previewImages: any;
   handleCancel: any;
   handleInputChange: any;
   handleSelectChange: any;
+  copyToOthers: (name: string, index: number) => void;
   selectOptions: any[];
 }) => (
   <>
@@ -23,10 +24,12 @@ const FileList = ({
           file,
           name,
           data,
+          storeKey
         }: {
           file: any;
           name: string;
           data: any | null;
+          storeKey: string;
         },
         index: any,
       ) => {
@@ -38,9 +41,10 @@ const FileList = ({
           index,
           name,
           selectOptions,
+          storeKey,
+          copyToOthers
         };
         const isImage = file.type === IMAGE;
-
         return (
           <Fragment key={file.id}>
             <div className="preview">
@@ -68,5 +72,4 @@ const FileList = ({
     )}
   </>
 );
-
 export default FileList;